Add files via upload
Group couldn't source original file - to make it easier to understand where data is from I copied the table from Power BI into a csv file  and removed some of the formulas we added to this table i.e.  Colour Intensity and Year, to make csv file as close to original as possible.